chaps: Increase the timeout of TestLogin.LoginNoPrivateWait to fix flakiness.
When the CPU is under heavy load, and it is not possible for PerformNonBlocking()
in chaps.cc to retry C_Login() a second time under 30ms, the test will fail.
This change increase the timeout to 1.5s so it should work. The reason why
increasing the timeout is chosen over mocking time is because:
1. Increasing timeout doesn't actually increase the run time for successful
cases. PerformNonBlocking() will return as soon as it finishes, and will not
wait for the full timeout.
2. PerformNonBlocking() uses base::TimeTick::Now() to poll for the current
time and see if deadline have passed. base::TimeTick::Now() is hard to mock
at the moment since internal::g_time_ticks_now_function is not exposed.

BUG=chromium:844114
TEST=chaps_test --gtest_repeat=5000000 --gtest_filter=TestLogin.LoginNoPrivateWait

Change-Id: I7e87a1c4b85d91396214fd6b3f4116d819038adf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2160570
Tested-by: John L Chen <zuan@chromium.org>
Commit-Queue: John L Chen <zuan@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>