chromeos-config: move all the test data to a test_data directory
There was a bunch of test data under libcros_config, which is pretty
sloppy. It's nice to keep this directory just the implementation of
libcros_config, especially since most of this test data was for
cros_config_host (not libcros_config) anyway.

So, all the test data was moved to a test_data directory, and the unit
tests were updated to point to the new location.

BUG=none
TEST=unit tests passing

Change-Id: I57a0d2d50640446c11ae1bb75895762022beb4a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2161905
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: C Shapiro <shapiroc@chromium.org>