shill: timeout_set: Add functionality for infinite lifetime elements
This change provides the ability for TimeoutSet to contain elements
with an infinite lifetime. An example of where this will be useful is
RDNSS and DNSSL, where ND_OPT_LIFETIME_INFINITY is used to indicate
that a particular set of entries have an infinite lifetime.

BUG=chromium:1033533
TEST=All unit tests are passing.

Change-Id: Idaf82a9c22d3024eda7bff82d906ce14713d83de
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2168332
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>