chromeos-config: allow for x86 identity info without smbios-name-match
smbios-name-match is optional now, so this is now an allowed identity
dictionary:

    {"identity":
      {"platform-name": "Dummy"}}

However, with a oneOf in the schema here, it now matches both the x86
and ARM groups. Changing this to anyOf allows the case where it
matches both groups.

Note: eventual plan is to move platform-name to /mosys or get rid of
it entirely (if we're lucky).

BUG=b:154129827
TEST=unit tests

Change-Id: Ifc66fb8d430b31cf739b784e5f0660f4d03f4e31
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2163360
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: C Shapiro <shapiroc@chromium.org>