debugd: Prevent |output_fd| from closing before the quipper process exits
The kernel closes fds before signaling process exit. This can lead to a
case where a user of GetPerfOutputFd() that reads from the output fd
until EOF and immedietly tries another GetPerfOutputFd() will see the
second request fail with the error "Existing perf tool running". This is
because debugd has not yet seen the signal that the perf process has
exited.

The solution here to to hold onto |output_fd| until debugd has seen the
process exit. This guarantees to the user of GetPerfOutputFd() that when
they see EOF from the output fd, it is safe to make another call to
GetPerfOutputFd().

BUG=b:150740987
TEST=Run 'test_that --no-retry DUT 'platform_DebugDaemonGetPerfOutputFd''
     100 times. Failure 'Existing perf tool running' should not be seen.

Change-Id: Ieec3245768371b269e524d221855b3eb3ea6f14d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2189990
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Chinglin Yu <chinglinyu@chromium.org>
Commit-Queue: Anand Mistry <amistry@chromium.org>
Tested-by: Anand Mistry <amistry@chromium.org>