webserver: Remove the Request ID post serving.
It is observed that for every request there is a mapping
of Request ID to ProtocolHandler is being added to the map
while processing the request. After successfully processing
the request, Request ID is not removed from the map causing
the heap memory to increase constantly. This CL fixes this
problem by removing the Request ID from the map once
it is successfully completed.

BUG=b:153568555
TEST=Compiled successfully and verified
     It is observed that post this change and testing for
     more than 16hours didn't observe significant increase
     in heap memory

Change-Id: Id6b201529a0136f222d22f41df675ba62eb48f2c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2162355
Reviewed-by: Raju Konduru <rkonduru@google.com>
Reviewed-by: John Riordan <jrio@chromium.org>
Reviewed-by: Kishan Kunduru <kkunduru@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Kishan Kunduru <kkunduru@chromium.org>
Tested-by: Kishan Kunduru <kkunduru@chromium.org>