dlcservice: Remove MOUNTABLE state from DlcState
MOUNTABLE state is not something that clients need to be aware of and
this state can confuse them. Also mountable is not a state that can be
tracked properly. Sometimes we migrate to this state sometimes not. Its
better to have an internal state variable to cache this state instead of
involving it in the states.

This CL also renames all instances of mountable to verified.

BUG=chromium:1074090
TEST=cros_workon_make --board reef --test dlcservice

Change-Id: I7676690f8e3d7e98b1e4b4b34f2cc4fd5f15f854
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2163915
Tested-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>