cryptohome: Add metrics for OOP mount duration.
In order to confidently extend out-of-process mounting to all session
types, add metrics to quantify how long the out-of-process mount
operation actually takes.

https://uma.googleplex.com/p/chrome/timeline_v2/?sid=99615dcce8d9e5ff43d281d3d33c6216
(Cryptohome.TimeToMountGuestAsync) tracks the duration of the entire
Guest session mount command (not just the OOP part). It suggests there
was no meaningful change at the 50th percentile, but there are some
spikes at the 95th percentile that we should probably dig into a little
bit. It wouldn't be unreasonable to think that the out-of-process
mechanism doesn't really have an effect on systems with little load,
but can make tail latency worse (since on a system under load, adding
the extra cost of launching a process could be noticeable.)

BUG=chromium:1074735
TEST=Log in Guest session, open chrome:histograms, see new histograms.

Change-Id: If709bb8484766e9e7f532d379d31a76e5074ccd6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2167212
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>
Commit-Queue: Jorge Lucangeli Obes <jorgelo@chromium.org>