shill: device: Stop tracking RDNSS lifetimes as lease expiration
CL:225278 introduced Device::TimeToNextDHCPLeaseRenewal for such
functionality as Wake-on-{W,}LAN, where the device would wake up prior
to DHCP lease expiration in order to request a lease renewal. In the
case of IPv6 without any sort of DHCPv6, there is no lease to be
renewed (perhaps expiration of entries in the router's neighbor cache
would be a concern if the magic packet was coming from outside the
LAN, but that is not necessarily the same as RDNSS lifetimes and
should be a case that we explicitly support if we want to go down that
road).

Note that TimeToNextDHCPLeaseRenewal still works for DHCPv6 as
DHCPConfig handles setting the lease expiration time, although that
code path is unused because it requires WiFi and DHCPv6 to be used
simultaneously.

BUG=chromium:1033533
TEST=All unit tests are passing.

Change-Id: I323512ba431e57b8b96caaa7efad6b8417c43d9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2168335
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>