chromeos-config: change autobrightness files to a single object
This was coded as an array, however from an email thread with the
authors, they never meant this.  The implementation in Chrome has not
landed yet, so not too late to fix this!

Since getting a system_file_v2 in an array was copy-pasted twice
already and we are doing the same thing, I left the campground a
little bit cleaner than I found it by making a common implementation
of _GetSystemFileV2.

BUG=chromium:1062070
TEST="cros_config_host get-autobrightness-files" for hatch

Cq-Depend: chromium:2168853
Change-Id: I7a3ef096efb4acf2e7afbc643503dca032755fda
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2168773
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Mengqi Guo <mqg@chromium.org>