hammerd: Use c_str() for fmap_find_area()
It looks like this takes a std::string but typically the underlying
string is a char pointer defined somewhere in the code, meaning it's NUL
terminated already. Using .data() means that in pre C++11 we won't
be guaranteed to send a NUL terminated string to the fmap_find_area()
function. Luckily we use a more modern version of C++ so .data() is
guaranteed to be NUL terminated. I had to look at this a couple times to
convince myself that there wasn't a problem here with non-NUL terminated
strings being passed around.

This is confusing and given that we're dealing with a C library used
from C++ let's use .c_str() so it's very explicit what the intent is.
This removes any doubt in the reader's mind about what is happening
here and clarifies the code. No functional changes.

BUG=None
TEST=Run hammerd_update_fw_load_ec_fuzzer and FEATURES="test"
emerge-amd64-generic hammerd

Change-Id: I880584f8e229cace88abc4e38123585de0a1314c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2172170
Tested-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Fei Shao <fshao@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>