cryptohome: Mockable KeyChallengeServiceFactory
Add a KeyChallengeServiceFactory interface with a real and a mocked
implementations, which allows to inject test instances of
KeyChallengeService into the Service class in tests.

This a pure refactoring CL that is a preparation for the implementation
of unit tests for challenge-response operations in Service.

BUG=chromium:1044558
TEST=cros_workon_make cryptohome --test

Change-Id: Ief563cd1997b003fc9a43149ea620a2d652cf8b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2172296
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>