hwsec-test-utils: |Decrypt| for |attestation::EncryptedData|.
This Cl implements a compound operations that decrypt the data in
|attestation::EncryptedData|.

The first consumer code will be
|VerifiedAccessChallenge::VerifyChallengeResponse| in the follow-up CL,
where we have to decrypt |KeyInfo|.

The |KeyDeriver| is used to dynamically inject the key derivation
function. Though in the hwsec-test-utils we don't expect the usecase of
other derivation in decryption case, considering portability potential
usage in unittests we reserve the feature of |KeyDeriver| as the
function input.

BUG=b:154672618
TEST=build ok.
TEST=work with the consumer code in the follow up CL

Change-Id: I234bf54336ddbc5572674a17a9d7990a9aa87ee8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2211627
Tested-by: Leo Lai <cylai@google.com>
Commit-Queue: Leo Lai <cylai@google.com>
Reviewed-by: Louis Collard <louiscollard@chromium.org>