power: Enable wake for the ACPI CREC device.
The EC is effectively enabled as a wake source if one of the devices
nested under it (e.g. a lid switch) is configured as a wake source.
Thus, if a device has a lid switch, then other EC host events that are
configured to wake the AP result in the AP waking. However, if no such
device is nested under the EC, those host events won't wake the AP.

Unconditionally enable wake for the CREC device to ensure host events
configured to wake the AP can wake the AP.

BUG=b:152913653
TEST=Suspend a puff to S0ix; run hostevent set 0x80000 in the ec console

Change-Id: Id6c1ae374739b609cd38dc3395b77d1b3835a02e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2172106
Tested-by: Sam McNally <sammc@chromium.org>
Commit-Queue: Sam McNally <sammc@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>