dlcservice: remove an unreachable return
We're trying to turn -Wunreachable-code-return on by default, and this
unreachable `return` statement was flagged. Removing it makes this code
clearer and easier to modify, since if a case is added to this `switch`
later, the user should get a -Werror'ed warning if they forget to add a
`return`.

BUG=chromium:1066998
TEST=CQ

Change-Id: Iee8b514943c41f2f5932c10f9ec5f6c2faf2e719
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2172347
Reviewed-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>