hwsec-test-utils: generate the serialized Google key proto
This CL creates the utility executable that dumps the default google
keys to be injected to attestationd, for testing purpose.

Currently it only dumps the default Google keys that are used for
production; gradually the keys will be altered then the local test infra
has the corresponding implementation.

BUG=b:155048010
TEST=build ok.
TEST=run the executable->restart attestationd->hwsec.Attestation

Cq-Depend: chromium:2175718
Change-Id: I45bb5b771d7acaf0b2cb7dd107980ab78e04dfd0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2175975
Tested-by: Leo Lai <cylai@google.com>
Commit-Queue: Leo Lai <cylai@google.com>
Reviewed-by: Louis Collard <louiscollard@chromium.org>