chromeos-config: validate only one mosys platform per-board
We have the goal to ultimately compile mosys for only a single
platform on unibuild boards.  Add validation to the schema that we
only use a single mosys platform in one config.

We learned with Deltaur that we should not be handing out too many
knobs to people on this.  This will prevent future projects from
making the same mistake.

BUG=chromium:1070692
TEST=unit tests passing

Cq-Depend: chromium:2171770
Change-Id: I894362e437d4de704d711d2bd70123096ac14f6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2173805
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: C Shapiro <shapiroc@chromium.org>