chromeos-config: add "cros_config_host get-mosys-platform" command
This is a command which will be used in the mosys ebuild to determine
the single platform to compile for.  Since the parent CL restricts the
config to using only one mosys platform, this will only ever print one
value, or exit with a failure status if no devices define a platform
name.

BUG=chromium:1070692
TEST=unit tests pass

Change-Id: I98454c6a5af03fcf82f9695b45952742570f873c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2173806
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: C Shapiro <shapiroc@chromium.org>