smbfs: Update inode's file path on rename
When a rename() occurs, the directory entries are changed, but the
underlying inode remains the same. Because libsmbclient operates on
file paths and not inodes, smbfs synthesizes inodes and maintains a
consistent mapping between indoes and paths (using InodeMap). So when a
rename happens, the renamed file's inode needs to have its path updated.

However, this presents a problem when the new path already exists. POSIX
says that the entry is overwritten (and potentially delete the existing
file at the new path). However, this requires temporarily having a
dangling inode, which is non-trivial to add. For now, disallow rename
where the new path already exists.

BUG=chromium:1075837
TEST=cros_run_unit_tests --board=chell --packages smbfs
TEST=On a mounted SMB share in Files app, create 5 new folders. Select
     all 5 folders (use ctrl-click) and drag-drop to another folder.
     Immediately navigate to the destination folder and notice all 5 new
     folders showing as folders (and not files).

Change-Id: Iacb485fa79f90230d8edaa4bc49ab08284777d22
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2174111
Tested-by: Anand Mistry <amistry@chromium.org>
Commit-Queue: Anand Mistry <amistry@chromium.org>
Reviewed-by: Sergei Datsenko <dats@chromium.org>