libipp: remove an unreachable return
We're trying to turn -Wunreachable-code-return on by default, and this
unreachable `return` statement was flagged. Removing
the unreachable return makes it clearer that the switch above always
returns. If someone does add a `break`'ing case down the line, they'll
get a -Werror'ed warning about lacking a `return;` with a valid value.

BUG=chromium:1066998, chromium:1077410
TEST=CQ

Exempt-From-Owner-Approval: Fix fuzzer builders

Change-Id: Ife9b6bd7669186ecc70e5858d6fc1bbc450c5d3e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2176616
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Manoj Gupta <manojgupta@chromium.org>