login: Allow LoginScreenStorage to store ephemeral data in lock screen
Previously, the LoginScreenStorage is not allowed to store any data
while there are active sessions to prevent a malicious session from
potentially manipulating data in the login screen.

This CL updates LoginScreenStorage so that is is allowed to store
ephemeral data while there are active sessions. Since the data will be
cleared when the session exits, the malicious session would not be able
to affect the login screen.

BUG=b:152380871
TEST=cros_run_unit_tests --board atlas --packages chromeos-login
TEST=Tested on device with chrome.loginScreenStorage API

Change-Id: I1bb2cbacfaeb8ae8cfb1b248708593bae644191d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2178574
Reviewed-by: Alexander Hendrich <hendrich@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>
Commit-Queue: Jit Yao Yap <jityao@google.com>
Tested-by: Jit Yao Yap <jityao@google.com>