shill: timeout_set: Remove use of EventDispatcher
Prior to this change, TimeoutSet relied on an EventDispatcher instance
to post tasks. This dependence led to the unfortunate situation
in which a TimeoutSet was not default constructible. Since this
dependence was only used for mocking in unit tests and those
expectations were not usefully covering much behavior, we remove this
dependence for the desirable property of default constructibility.

BUG=chromium:1033533, chromium:1019834
TEST=All unit tests are passing.

Change-Id: Ie1b92c45c14d2b885d28b666dab9644700fba394
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2180971
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>