hermes: Add confirmation_code parameters for Install* methods
An SM-DP+ may demand a confirmation code from the user prior to
allowing a profile download to proceed. Note that this may be the case
even if the activation code doesn't indicate a required confirmation
code, or a profile download is being triggered using an SM-DS event
record.

If a profile download is attempted without a confirmation code but the
SM-DP+ requires one, the Install* method will fail with
kErrorNeedConfirmationCode.

BUG=chromium:1054563
TEST=All unit tests are passing.

Change-Id: I7a7b0f35d7ebaf5129e3b8050701893e1e5efcd8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2185481
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>