dlcservice: Hash DLC images for Install + Reboot Fix
Squashing two bugs with one CL as the flow is closely related.

When a DLC image is requested to be installed and the DLC image happens
to be present on the disk, dlcservice will hash the current active slot
of the DLC image to verify if it is a match with the rootfs imageloader
manifest hash.

This also fixes the currently broken cros deploy workflow when working
with DLC packages/ebuilds.

This also fixes the issue on reboot that deletes verified images
transitioning from |NOT_INSTALLED| to |INSTALLED| state.

BUG=chromium:153397275
TEST=FEATURES=test emerge-$B dlcservice

Change-Id: If624f4140b96ff6a0dc11d225ccdb160db08c269
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2188970
Tested-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Jae Hoon Kim <kimjae@chromium.org>