shill: crypto: Unify CryptoProvider and CryptoRot47 code
Prior to this change, CryptoProvider provided backwards-compatibility
and migration support for multiple crypto modules (CryptoInterface
children). After removing the CryptoDecCbc module, however, the only
used crypto module was CryptoRot47.

This change unifies the CryptoProvider and CryptoRot47 code, with the
intention that if a new crypto module is needed in the future, this
change can be reverted. This is done so that the current code can
benefit from increased simplicity while future code can still use the
previous flexibility.

BUG=None
TEST=All unit tests are passing.

Change-Id: I828e4860e2e557932009d94162d7df0000542b56
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2197540
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>