smbfs: Introduce 'weak' inodes and use for readdir()
In FUSE, readdir() returns an inode number for each directory entry, but
must not increment the refcount for that inode. This is because the
st_ino field field of a directory entry passes through the kernel
to userspace without being processed in any way.

In order to cache inode attributes from libsmbclient's readdirplus(),
inodes need to allocated so that they're consistently returned on later
calls to lookup(). However, these inodes can't have a non-zero refcount
and can't exist for any other purpose until lookup() is called. For
this, weak inodes can be used.

BUG=chromium:1054711
TEST=Add SMB share in Files app and navigate to folder containing 10,000
     files. Listing folder should be fast (couple of seconds).

Change-Id: I481acb779c9cc096cf62911680c9314fb477e327
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2198337
Reviewed-by: Sergei Datsenko <dats@chromium.org>
Tested-by: Anand Mistry <amistry@chromium.org>
Commit-Queue: Josh Simmons <simmonsjosh@google.com>