remove MessageLoopIdle and its usage
MessageLoopIdle is used to wait for a MessageLoop to be idle in
libhwsec. And it's only used by AttestationServiceTest to wait for
base::Thread::message_loop() to be idle.

It's strongly discouraged to synchronize with base::Thread by hacking
with the MessageLoop, thus the MessageLoop will not be exposed at next
version of libchrome.

base::Thread::FlushForTesting doesn't fit the requirement of this
unittest since it only wait for current PendingTasks, but not the
RunLoop to be idle. So Add a OnceCallback for
AttestationService::Iniitalize/IniitalizeTask/PrepareForEnrollment so
the service can call a callback when PostTask posted tasks complete.

BUG=chromium:1066447
TEST=emerge and unittest for libhwsec, attestation.

Change-Id: Ib47fb0a9e416ffae7c52e2a813d266845324e05f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2198881
Tested-by: Qijiang Fan <fqj@google.com>
Commit-Queue: Qijiang Fan <fqj@google.com>
Reviewed-by: Leo Lai <cylai@google.com>