webserver: Fix for webservd crash
It is observed that webservd is getting crashed when it fails to
serve request because of "Internal Server Error". It calls OnError()
function upon failure case and observed that Request instance is
NULL which leads to crash. Logged the error message and returning if
the Request instance is invalid.

BUG=b:140407692
TEST=cros_workon_make --board=gale --test webserver

Change-Id: Ia13da3c8fd79430a1809e417754dcee344af322a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2198884
Tested-by: SHAMANTHA KUMAR PERAMACHANAHALLI RAMARAO <shamanthakumar@google.com>
Tested-by: Rakesh K <k.rakesh@globaledgesoftware.corp-partner.google.com>
Commit-Queue: Kishan Kunduru <kkunduru@chromium.org>
Reviewed-by: Amrut Shetty <amruts@google.com>
Reviewed-by: Kishan Kunduru <kkunduru@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>