vm_tools: Add support for privileged containers
This change plumbs an optional privileged field in the requests involved
in starting a container. This will enable clients to start privileged
containers when allowed by the host. The crostini client is also
modified to support this field. A TAST test will be added in a separate
commit.

BUG=chromium:1072784
TEST=None

Change-Id: I92f69d24a2239de22007c0803110cc86d7af2837
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2198691
Tested-by: Abhishek Bhardwaj <abhishekbh@chromium.org>
Commit-Queue: Abhishek Bhardwaj <abhishekbh@chromium.org>
Auto-Submit: Abhishek Bhardwaj <abhishekbh@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>