vm_tools: vmc: guess the CROS_USER_ID_HASH value
When there is only one user on a device there is no reason to require
the user to set the CROS_USER_ID_HASH variable.

This removes a small annoyance in the development/test workflow where
this variable is not set when you ssh to the device as root.

BUG=None
TEST=compiled and ran on eve device.

Change-Id: I46e98780d319bb99c096fb82f02ed2d66ff4064a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2206116
Tested-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Matthew Warton <mwarton@google.com>
Reviewed-by: Stephen Barber <smbarber@chromium.org>