nnapi: Compile more files and tests from aosp/system/core/base
Adds in more files from system/core/base that will compile without
liblog at least having a partial implementation.

Moves the definition of the test runner to be closer to the definition
of the library to make it easier to compare between the two.

BUG=b:156714501
TEST=FEATURES="test" emerge-octopus chromeos-base/nnapi

Change-Id: I357773b8a0c5ead7d7ffe600d56a56d8e19d927e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2206571
Commit-Queue: Stuart Langley <slangley@chromium.org>
Tested-by: Stuart Langley <slangley@chromium.org>
Reviewed-by: Alan Green <avg@chromium.org>