hermes: Have SM-DS create pending Profile object rather than Event tuple
Prior to this change, an SM-DS request would return a list of Event
Record tuples, which clients could then pass to an Install method in
order to download the Profile from the corresponding SM-DP+. One major
downside of this interface is that Profile metadata would not be
available to the client prior to calling the Install method.

Since the Profile metadata contains iccid--which Profile DBus objects
use for their paths--we can instead have an SM-DS request both get
Event Records *and* request metadata from the SM-DP+, resulting in a
Profile object in a pending state.

Note that both the functionality for getting Event Records and for
retrieving just Profile metadata will need to be added to the
google-lpa library before this feature is actually available to
clients.

BUG=chromium:847619, chromium:1054563
TEST=All unit tests are passing.

Change-Id: I30af8e4e33037ae00df4b3a28047c2fc52eada4c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2207786
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>