cros-disks: Use MS_NOSYMFOLLOW flag when available
In our 5.4 kernel, the mount syscall accepts a new MS_NOSYMFOLLOW flag
that replaces the existing string nosymfollow option.  When parsing the
nosymfollow mount option, we need to translate it to the new flag.

The mount syscall doesn't provide a way to detect at runtime which flags
are supported, so we need to pass the flag both ways for cross-kernel
compatibility.  This is safe because old kernels will ignore the
MS_NOSYMFOLLOW flag and new kernels will ignore the string option.  The
string option can be removed once all devices have moved to a kernel
that includes MS_NOSYMFOLLOW.

BUG=b:152074038
TEST=test_that -b {zork,hatch} platform_CrosDisks*

Change-Id: Idfbb6625e51fcd56a93dfb47b34f1c97541fd62c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2209731
Reviewed-by: Sergei Datsenko <dats@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>