arc: Bind mount /vendor/build.prop too
In addition to /default.prop and /system/build.prop, ARCVM also
modifies /vendor/build.prop at runtime and mount the modified one in
the guest. This CL is to do the same for ARC container.

In P, mounting the vendor file does not really change the container's
properties. This is just to make ARC container's property loading
steps a little closer to ARCVM's.

BUG=b:156315252
TEST=run arc.Boot, then verify 'ls -lZ /vendor/build.prop' returns
  u:object_r:vendor_file:s0, 'getprop ro.vendor.build.date' returns
  non-empty string.

Change-Id: Id45a8d97571325a39ef176a98f1c83c24109f588
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2211520
Tested-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Ereth McKnight-MacNeil <ereth@chromium.org>
Commit-Queue: Yusuke Sato <yusukes@chromium.org>