nnapi: Temporarily add required AOSP logging headers.
Due to crbug.com/1083354 we are unable to get the mirror of the liblog
git repository setup to access the header files.

This CL temporarily adds the headers files that are needed by
android-base/logging.h.

They are being landed as a separate patch so we can revert it once the
gsubtree bug is resolved.

Taken from AOSP commit 4ea509e2.

BUG=b:156714501
TEST=None

Change-Id: Iba656647f00927afaf707718a959a6bd8fd7dbce
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2211117
Reviewed-by: Alan Green <avg@chromium.org>
Commit-Queue: Stuart Langley <slangley@chromium.org>
Tested-by: Stuart Langley <slangley@chromium.org>