dlcservice: Simplify the installation logic
This CL is addressing a few related issues.

Currently, there are a few different routes for installing internally
and it has become hard to manage these routes and add new features so we
need to see if we can merge them to a single point of contact. For
example we had two sets of codes that try to verify and mount an
image. One in SetupInitInstall() and in FinishInstall(). But since we
can (and should) mount an image that was already installed, we can merge
all different codes into one. This breaks everything into two functions
of DlcBase:

- Install: This function tries to install an image either through
preloading or mounting it if the image was already verified.

- FinishInstall(): This function is called for two purposes, one for
finishing an install that was being done by the update_engine. And one
is called by the Install() function after it has enough knowledge that
the image on disk is healthy and ready to be mounted.

This CL also fixes an issue where a NOT_INSTALLED DLC is simply deleted
without unmounting. But the image can be mounted on a previous session
of dlcservice.

This CL fixes another issue when the install would've been failed only
based on the status of the update_engine without regard to whether we
need the update_engine for any installation of not. This CL first tries
to install the DLC with other means, and only if was not successful by
those means, it will try to install through update_engine. The drawback
is that if the update_engine is not ready, we will waste an image
creation and deletion. But I put a TODO to optimize this better.

Another issue this fixes is we used to set the active flags at the
beginning of install and remove if if the install failed. But there is
no point to this if we are to remove these flags immediately because
only during an update these flags are needed. So in this CL we just set
these active flags only and only if the installation has been
successfully finished. This resulted to removal of a lot of
SetDlcActiveValue mocks from unrelated unittests.

BUG=chromium:1074173
TEST=cros_workon_make --board reef --test dlcservice
TEST=dlcservice_util --install dummy-dlc
TEST=tast run chromeos6-row4-rack9-host15.cros platform.DLCServic

Change-Id: I90ea9d329a96f7544e4118765c39a8c21389da3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2215138
Tested-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>