dlcservice: Add size for HashFile and CopyAndHashFile
These two functions should not rely on the source image size, they
should rely on the size indicated in the manifest instead. For example,
DLC images have a preallocated portion at the end and we don't want to
hash those because the manifest hash doesn't contain those bytes.

Another problem is that the unittests are assuming the file filled with
zero is the actual file. But this can cause problems in the future as it
has done now for these two functions working properly even though they
shouldn't have. This CL fills the correct file content to character '1'
and changes the corresponding hashes in the manifest files.

BUG=chromium:1086309
TEST=cros_workon_make --board reef --test dlcservice

Change-Id: I8062be019eb723b20318f34d70404fad7188d642
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2216233
Tested-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>