system-proxy: Improve worker_config_fuzzer exit
Worker processes are configured by sending serialized protos to the
standard input.

This CL replaces the fuzzer test exit condition so that a callback
is called that will quit the current message loop after the input is
read and processed instead of waiting for the task queue to be empty.

BUG=chromium:1086346
TEST=ran the fuzzer test for over an hour

Change-Id: I869345054f09e3224d1a02aaf7160faa0ad461b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2215171
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Tested-by: Andreea-Elena Costinas <acostinas@google.com>
Commit-Queue: Andreea-Elena Costinas <acostinas@google.com>