dlcservice: Don't fail the install when failing to write the verified pref
We should not really failing an install if we fail to write the verified
pref. We have done a lot of work and the image seems to be verified. So
we should not dismiss all that effort just because we temporarily
couldn't write a small file. This CL keeps the image as verified in the
current session. On the next sessions the image can be downloaded again
or reverified. Currently this should happen anyway.

BUG=None
TEST=cros_workon_make --board reef --test
TEST=dlcservice_client --install dummy-dlc

Change-Id: I130aa54969f88ba8532dcdbe318ee3caf3b5419f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2222793
Tested-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>