init: Add service to clean up after early-failure.
The early-failure service (which is used to test handling of early-boot
crashes) generates a crash dump every time a test image reboots.  Since
it runs during early boot, it is not possible for it to examine any
stateful files to determine if crash.EarlyCrash is actually in progress,
so it crashes every boot.

These extra crash dumps create noise in crash-monitoring dashboards and
waste disk space.

To resolve this, add a new service that removes crashes that the
early-failure service creates after any pending crash tests finish.

BUG=chromium:1039378
TEST=tast run $DUT crash.EarlyCrash.* crash.KernelCrash.*
TEST=reboot; check that /var/spool/crash doesn't have a coreutils crash
TEST=don't delete crash-test-in-progress and verify cleanup times out \
     and leaves coreutils crash

Change-Id: Ied87b1eb6cd54705461b579f8adae8b7c3e2f228
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2222791
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>