shill: wifi: update Security when we see an endpoint
WiFiService::UpdateFromEndpoints() appears to be updating the Security
property, but it doesn't actually do that. Thus, until we actually
connect to an endpoint (and so we have current_endpoint_ set, and
WiFiService::GetSecurity() gives a more specific answer), the Security
property will always be generic (e.g., "psk" instead of "wpa", "rsn",
etc.).

This has the following effects, for psk networks:
 * D-Bus: Security property always remains "psk" for unconnected
   networks, instead of more specific answers based on visible BSSs
 * Service::SecurityLevel() can't account for WPA vs. WPA2 properly

This fixes those:
 * D-Bus will show correctly, as soon as we see a BSS on the air
 * SecurityLevel() will be accurate, whether we're connected or not

I kill WiFiServiceTest.SecurityFromCurrentEndpoint along the way,
because it's testing illegal behavior (we never should see a service get
Connected to a mismatching SecurityClass) that we're now DCHECK()'ing.
Implement a similar test in WiFiProviderTest instead.

BUG=b:157935328
TEST=unit tests; `connectivity show Services`, see non-connected
     services have appropriate "wpa" or "rsn" values

Change-Id: If10624bc06201de08cafa92c59c45f7e27274aad
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2238461
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>